Added tests and alternative algorithm for ordering
The new algorithm doesn't build a dependency graph. Instead if builds the list
of nodes with dependencies that are met based on the assumption that there will
always be at least one node that has no dependencies or all of its dependencies
already met.

The algorithm passes all existing tests, as well as the new test that revealed
a problem with the previous algorithm.

Admittedly, implementation may be quite inefficient, as am I at 2:30am. Further
refinement is probably needed.

Signed-off-by: Branko Vukelic <branko@outernet.is>