Fix for failure to account for met deps
When code tests for unmet dependencies that are not in the remaining
dependencies, it fails to take into account dependencies that are already met.
For example, if node A depends on B and C, and C is already satisfied, it will
test whether both B and C and in the remained (only B), and declare unmet
dependency. This patch fixes this issue.

Signed-off-by: Branko Vukelic <branko@outernet.is>